pyproject.toml
[project]
name = "bilibili-video-maker"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.5",
    "playwright>=1.40.0",
    "black>=24.0.0",
    "loguru>=0.7.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "ffmpeg-python>=0.2.0",
    "openai>=1.0.0",
    "bilibili-api-python>=17.4.1",
    "apscheduler>=3.10.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
